Use Long for mdms-v2 count contract to match PostgreSQL BIGINT
COUNT(*) returns bigint in Postgres; mapping it to Integer would throw
on counts above 2,147,483,647. Addresses CodeRabbit review comment on
PR #1405.

+++ core-services/mdms-v2/src/main/java/org/egov/infra/mdms/controller/MDMSControllerV2.java
@@ -54,7 +54,7 @@ public class MDMSControllerV2 {
      */
     @RequestMapping(value="_count", method = RequestMethod.POST)
     public ResponseEntity<MdmsCountResponseV2> count(@Valid @RequestBody MdmsCriteriaReqV2 masterDataSearchCriteria) {
-        Integer totalCount = mdmsServiceV2.count(masterDataSearchCriteria);
+        Long totalCount = mdmsServiceV2.count(masterDataSearchCriteria);
         return new ResponseEntity<>(ResponseUtil.getMasterDataV2CountResponse(RequestInfo.builder().build(), totalCount), HttpStatus.OK);
     }
