+++ core-services/mdms-v2/src/main/java/org/egov/infra/mdms/model/MdmsCountResponseV2.java
@@ -20,5 +20,5 @@ public class MdmsCountResponseV2 {
     private ResponseInfo responseInfo = null;
 
     @JsonProperty("totalCount")
-    private Integer totalCount = null;
+    private Long totalCount = null;
 }
